requirements.txt
numpy>=1.19.0
faiss-cpu>=1.7.0 
matplotlib
